Sharing Done , preview done

+++ app/Http/Controllers/FilesController.php
@@ -7,6 +7,7 @@ use App\User;
 use Auth;
 use App\Files;
 use App\Trash;
+use App\Shared;
 use File;
 use Storage;
 use Response;
@@ -23,7 +24,7 @@ class FilesController extends Controller
     }
     public function index()
     {
-        $files = Files::orderBy('created_at','desc')->paginate(10);
+        $files = Files::orderBy('created_at','desc')->where('user_id',Auth::id())->paginate(10);
 
         return view('files.myfiles')->with('files',$files);                                                                                                           
     }
@@ -40,24 +41,36 @@ class FilesController extends Controller
     public function getDownload($id)
     {   
         $file = Files::find($id);
+        $currentuser = Auth::id();
         $file_name = $file->file;
         $user_id=$file->user_id;
-        $path = storage_path().'\\app'."\\".Auth::id().'\\files'.'\\'.$file_name;
-
+        $path = storage_path().'\\app'."\\".$user_id.'\\files'.'\\'.$file_name;
+        $shared = Shared::where('file_shared',$file_name)->where('owner',$user_id)->first();
         /* ::where('status' , 0)
      ->where(function($q) {
          $q->where('type', 'private')
            ->orWhere('type', 'public');
      })
      ->get(); */
-        if($user_id == Auth::id())
+        if($user_id == Auth::id() )
         {
                 if (file_exists($path) ) 
                 {
                     return Response::download($path);
                 }
         }
-        return redirect('/')->with('error','Access Denied');
+        else if( Auth::id() == $shared->shared_with ) { 
+            
+            if (file_exists($path) ) 
+            {
+                
+                return Response::download($path);
+            }
+        }
+        else{
+            return redirect('/')->with('error',"Access Denied");
+        }
+        
     }
     /**
      * Store a newly created resource in storage.
@@ -137,14 +150,18 @@ class FilesController extends Controller
     {
         $filename = Trash::find($id);
         $delete = Storage::delete(Auth::id().'/trash'.'/'.$filename->file);
-        $filename->delete();
-        return back()->with('success','Deleted permanently');
+        if(Auth::id() == $filename->user_id)
+        {
+            $filename->delete();
+            return back()->with('success','Deleted permanently');
+        }
+        
     }
     public function m2t($id)
     {
         $filename = Files::find($id);
         $move = Storage::move(Auth::id().'/files'.'/'.$filename->file,Auth::id().'/trash'.'/'.$filename->file);
-        if(Storage::exists(Auth::id().'/trash'.'/'.$filename->file))
+        if(Auth::id()==$filename->user_id && Storage::exists(Auth::id().'/trash'.'/'.$filename->file))
             {
                 $filename->delete();
                 $file = new Trash;
@@ -152,16 +169,16 @@ class FilesController extends Controller
                 $file->file = $filename->file;
                 $file->type = $filename->type;
                 $file->save();
-
+                return back()->with('success','Moved to trash');
             }
-        return back()->with('error','Moved to trash');
+        return back()->with('error','Access Denied');
     }
 
     public function m2f($id)
     {
         $filename = Trash::find($id);
         $move = Storage::move(Auth::id().'/trash'.'/'.$filename->file,Auth::id().'/files'.'/'.$filename->file);
-        if(Storage::exists(Auth::id().'/files'.'/'.$filename->file))
+        if(Auth::id()==$filename->user_id && Storage::exists(Auth::id().'/files'.'/'.$filename->file))
             {
                 $filename->delete();
                 $file = new Files;
@@ -169,14 +186,14 @@ class FilesController extends Controller
                 $file->file = $filename->file;
                 $file->type = $filename->type;
                 $file->save();
-
+                return back()->with('success','Moved to My Files');
             }
-        return back()->with('error','Moved to My Files');
+        return back()->with('error','Access Denied');
     }
 
     public function trash()
     {
-        $trash = Trash::orderBy('created_at','desc')->paginate(10);
+        $trash = Trash::orderBy('created_at','desc')->where('user_id',Auth::id())->paginate(10);
         return view('files.trash')->with('trash',$trash);
     }
 
@@ -193,12 +210,37 @@ class FilesController extends Controller
         return Response($img);
         }
         } */
-    public function getImage($filename) {
-        $path = '/var/www/InAir/app/1'.'/files'.'/'.$filename;
-        $type = "image/jpeg";
-        header('Content-Type:'.$type);
-        header('Content-Length: ' . filesize($path));
-        readfile($path);
+    public function getPrev($id) {
+        $file = Files::find($id);
+        $currentuser = Auth::id();
+        $file_name = $file->file;
+        $user_id=$file->user_id;
+        $path = storage_path().'\\app'."\\".$user_id.'\\files'.'\\'.$file_name;
+        $shared = Shared::where('file_shared',$file_name)->where('owner',$user_id)->first();
+        /* ::where('status' , 0)
+     ->where(function($q) {
+         $q->where('type', 'private')
+           ->orWhere('type', 'public');
+     })
+     ->get(); */
+        if($user_id == Auth::id() )
+        {
+                if (file_exists($path) ) 
+                {
+                    return Response::file($path);
+                }
+        }
+        else if( Auth::id() == $shared->shared_with ) { 
+            
+            if (file_exists($path) ) 
+            {
+                
+                return Response::file($path);
+            }
+        }
+        else{
+            return redirect('/')->with('error',"Access Denied");
+        }
     
         }
 }
